Update readme.md for mongo shell command (#22)
Replace "mongo" with "mongosh" command.

As in the course it's used always the latest mongo version:
"The legacy mongo shell was deprecated in MongoDB 5.0 and removed in MongoDB 6.0 in favor of the new MongoDB Shell, mongosh. Some legacy methods are unavailable or have been replaced with updated methods in mongosh. To maintain backwards compatibility, the legacy methods that mongosh supports use the same syntax as the corresponding methods in the mongo shell."

Reference: https://www.mongodb.com/docs/mongodb-shell/